backend/requirements.txt
# Web Framework
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.22

# Computer Vision & ML
opencv-python-headless==4.13.0.90
ultralytics==8.4.7
torch==2.9.1
torchvision==0.24.1

# Data Processing
numpy==2.3.2
pandas==2.3.3
joblib==1.5.3
scikit-learn==1.6.1
pillow==11.3.0

# AI APIs
openai==2.21.0
groq==1.0.0

# Database
pymongo==4.7.3

# Utilities
python-dotenv==1.2.1
requests==2.32.5